Moved CallProvider
